requirements.txt
click>=8.1.7
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
colorama>=0.4.6
rich>=13.7.0
openai>=1.6.0
fastapi>=0.108.0
uvicorn>=0.25.0
google-api-python-client>=2.123.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
markitdown[all]
requests>=2.31.0